Some minor to align with Conda recipes
We have proposed here two minor changes in the specification to make the Dockerfile's more complaint with conda recipes. The website field is replaced by home and software.version is replaced by software_version. Also, base.image is replaced by base_image. 

As a recommendation for future fields that are named with more than one word the *_* should be used.